frontend: Extract ListItemLink and SidebarItem from Sidebar
