+++ frontend/src/components/Sidebar/Sidebar.tsx
@@ -1,23 +1,16 @@
 import { Icon } from '@iconify/react';
-import { Tooltip } from '@material-ui/core';
 import Box from '@material-ui/core/Box';
 import Button from '@material-ui/core/Button';
-import Collapse from '@material-ui/core/Collapse';
 import Drawer from '@material-ui/core/Drawer';
 import Grid from '@material-ui/core/Grid';
 import List from '@material-ui/core/List';
-import ListItem, { ListItemProps } from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import ListItemText from '@material-ui/core/ListItemText';
 import { makeStyles, withStyles } from '@material-ui/core/styles';
 import SvgIcon from '@material-ui/core/SvgIcon';
 import clsx from 'clsx';
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { generatePath, useHistory } from 'react-router';
-import { Link as RouterLink, LinkProps as RouterLinkProps, useLocation } from 'react-router-dom';
-import { createRouteURL, getRoute } from '../../lib/router';
-import { getCluster, getClusterPrefixedPath } from '../../lib/util';
+import { useLocation } from 'react-router-dom';
 import { setSidebarSelected, setWhetherSidebarOpen } from '../../redux/actions/actions';
 import { useTypedSelector } from '../../redux/reducers/reducers';
 import { SidebarEntry } from '../../redux/reducers/ui';
@@ -27,6 +20,7 @@ import { ReactComponent as LogoWithTextLight } from '../../resources/logo-light.
 import CreateButton from '../common/Resource/CreateButton';
 import NavigationTabs from './NavigationTabs';
 import prepareRoutes from './prepareRoutes';
+import SidebarItem from './SidebarItem';
 import VersionButton from './VersionButton';
 
 export const drawerWidth = 330;
@@ -79,66 +73,6 @@ const useStyle = makeStyles(theme => ({
     width: 'auto',
   },
 }));
-
-const IconTooltip = withStyles(theme => ({
-  tooltip: {
-    backgroundColor: '#474747',
-    color: '#fff',
-    minWidth: 60,
-    padding: '0.5rem',
-    fontSize: '0.8rem',
-    border: '1px solid #474747',
-    marginLeft: '1rem',
-  },
-}))(Tooltip);
-
-interface ListItemLinkProps {
-  primary: string;
-  to: string;
-  name: string;
-  icon?: object;
-}
-
-function ListItemLink(props: ListItemLinkProps) {
-  const { primary, to, icon, name, ...other } = props;
-  const namespaces = useTypedSelector(state => state.filter.namespaces);
-  const namespaceURLString = namespaces.size !== 0 ? `?namespace=${[...namespaces].join('+')}` : '';
-
-  const renderLink = React.useMemo(
-    () =>
-      React.forwardRef<any, Omit<RouterLinkProps, 'to'>>((itemProps, ref) => (
-        <RouterLink to={{ pathname: to, search: namespaceURLString }} ref={ref} {...itemProps} />
-      )),
-    [to, namespaceURLString]
-  );
-  let listItemLink = null;
-
-  if (icon) {
-    listItemLink = (
-      <ListItemIcon>
-        <Icon icon={icon} width={30} height={30} />
-      </ListItemIcon>
-    );
-  }
-
-  let listItemLinkContainer = listItemLink;
-  if (!primary) {
-    listItemLinkContainer = listItemLink && (
-      <IconTooltip title={name} placement="right-start">
-        {listItemLink}
-      </IconTooltip>
-    );
-  }
-
-  return (
-    <li>
-      <ListItem button component={renderLink} {...other}>
-        {listItemLinkContainer}
-        <ListItemText primary={primary} />
-      </ListItem>
-    </li>
-  );
-}
 
 export default function Sidebar() {
   const sidebar = useTypedSelector(state => state.ui.sidebar);
@@ -206,144 +140,6 @@ export default function Sidebar() {
         </Grid>
       </Grid>
     </Drawer>
-  );
-}
-
-const useItemStyle = makeStyles(theme => ({
-  nested: {
-    '& .MuiListItem-root': {
-      paddingLeft: theme.spacing(9),
-    },
-  },
-  linkMain: {
-    textTransform: 'uppercase',
-  },
-  link: {
-    color: theme.palette.sidebarLink.main,
-    '&:hover': {
-      color: theme.palette.primary.contrastText,
-      backgroundColor: `${theme.palette.sidebarLink.selectedBg}!important`,
-    },
-    '& svg': {
-      color: theme.palette.sidebarLink.main,
-    },
-    '& *': {
-      fontSize: '1.2rem',
-    },
-    '& .MuiListItemIcon-root': {
-      minWidth: 0,
-      marginRight: '12px',
-    },
-    paddingLeft: '23px',
-  },
-  linkSelected: {
-    color: theme.palette.primary.contrastText,
-    // For some reason we need to use important as even though we override the
-    // "selected" class, it uses the default one first.
-    backgroundColor: `${theme.palette.sidebarLink.selectedBg}!important`,
-    '& *': {
-      fontWeight: 'bold',
-    },
-  },
-  linkActive: {
-    color: theme.palette.primary.contrastText,
-    '&:hover': {
-      backgroundColor: theme.palette.sidebarLink.selectedBg,
-      svg: {
-        color: theme.palette.primary.contrastText,
-      },
-    },
-    '& svg': {
-      color: theme.palette.primary.contrastText,
-    },
-    '& *': {
-      fontSize: '1.2rem',
-    },
-  },
-  linkSmallWidth: {
-    backgroundColor: 'inherit !important',
-  },
-}));
-
-interface SidebarItemProps extends ListItemProps, SidebarEntry {
-  selectedName?: string | null;
-  hasParent?: boolean;
-  fullWidth?: boolean;
-}
-
-function SidebarItem(props: SidebarItemProps) {
-  const {
-    label,
-    name,
-    url = null,
-    useClusterURL = false,
-    subList = [],
-    selectedName,
-    hasParent = false,
-    icon,
-    fullWidth = true,
-    ...other
-  } = props;
-
-  const classes = useItemStyle({ fullWidth });
-
-  let fullURL = url;
-  if (fullURL && useClusterURL && getCluster()) {
-    fullURL = generatePath(getClusterPrefixedPath(url), { cluster: getCluster()! });
-  }
-
-  if (!fullURL) {
-    let routeName = name;
-    if (!getRoute(name)) {
-      routeName = subList.length > 0 ? subList[0].name : '';
-    }
-    fullURL = createRouteURL(routeName);
-  }
-
-  function isSelected() {
-    return name === selectedName;
-  }
-
-  function shouldExpand() {
-    return isSelected() || !!subList.find(item => item.name === selectedName);
-  }
-
-  const expanded = subList.length > 0 && shouldExpand();
-  let linkClass = classes.link;
-  if (hasParent || expanded) {
-    linkClass += ' ' + classes.linkActive;
-  }
-  if (!hasParent) {
-    linkClass += ' ' + classes.linkMain;
-  }
-  if (!fullWidth) {
-    linkClass += ' ' + classes.linkSmallWidth;
-  }
-
-  return (
-    <React.Fragment>
-      <ListItemLink
-        selected={isSelected()}
-        to={fullURL || ''}
-        primary={fullWidth ? label : ''}
-        classes={{
-          selected: classes.linkSelected,
-        }}
-        className={linkClass}
-        icon={icon}
-        name={label}
-        {...other}
-      />
-      {subList.length > 0 && (
-        <Collapse in={fullWidth && expanded}>
-          <List component="div" disablePadding className={classes.nested}>
-            {subList.map((item, i) => (
-              <SidebarItem key={i} selectedName={selectedName} hasParent {...item} />
-            ))}
-          </List>
-        </Collapse>
-      )}
-    </React.Fragment>
   );
 }
